a_zip_file_segments/rick_system_zip/openalgo-openalgo-theme/mbot-master/requirements.txt
aioboto3>=12.0.0
aiobotocore>=2.7.0
aiohttp>=3.9.0b0
aioitertools>=0.11.0
aiosignal>=1.3.1
annotated-types>=0.6.0
anyio>=3.7.1
async-timeout>=4.0.3
attrs>=23.1.0
boto3>=1.28.64
botocore>=1.31.64
certifi>=2023.7.22
cffi>=1.16.0
chardet>=5.2.0
charset-normalizer>=3.3.1
click>=8.1.7
cryptography>=41.0.5
distro>=1.8.0
dnspython>=2.4.2
docutils>=0.20.1
ecdsa>=0.18.0
email-validator>=2.1.0.post1
fastapi>=0.104.0
frozenlist>=1.4.0
h11>=0.14.0
httpcore>=1.0.0
httptools>=0.6.1
httpx>=0.25.0
idna>=3.4
itsdangerous>=2.1.2
Jinja2>=3.1.2
jmespath>=1.0.1
MarkupSafe>=2.1.3
multidict>=6.0.4
openai>=1.0.0b3
orjson>=3.9.10
packaging>=23.2
pyasn1>=0.5.0
pycparser>=2.21
pydantic>=2.4.2
pydantic-extra-types>=2.1.0
pydantic-settings>=2.0.3
pydantic_core>=2.11.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
python-jose>=3.3.0
python-multipart>=0.0.6
PyYAML>=6.0.1
requests>=2.31.0
rsa>=4.9
s3transfer>=0.7.0
six>=1.16.0
sniffio>=1.3.0
starlette>=0.31.1
tqdm>=4.66.1
types-pyasn1>=0.6.0.20250208
types-python-jose>=3.4.0.20250224
types-requests>=2.32.0.20250306
typing_extensions>=4.8.0
ujson>=5.8.0
urllib3>=2.0.7
uvicorn>=0.23.2
uvloop>=0.19.0
watchfiles>=0.21.0
websockets>=12.0
wrapt>=1.15.0
yarl>=1.9.2
